requirements.txt
pyrogram==2.0.106
tgcrypto
python-dotenv
motor
pymongo
pillow
ffmpeg-python
yt-dlp
aiohttp
beautifulsoup4
requests
cryptography
pyqrcode
pngcanvas
